pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cmc-csci040-valerie"
version = "0.1.3"
description = "A pirate chatbot"
readme = "README.md"
dependencies = [
    "groq",
    "python-dotenv",
]

[project.scripts]
chat = "chat:repl"
